Rename `mars_rover.py` -> `rover.py`
This is for simplicity
