+++ tests/integration/test_mars_rover.py
@@ -1,7 +1,7 @@
 import pytest
 
 from location import CardinalPoint, Position
-from mars_rover import Rover
+from rover import Rover
 from rover_api import RoverAPI
 from tests.helpers import assert_rover_state
