+++ tests/units/test_mars_rover_forward.py
@@ -4,7 +4,7 @@ import mars
 import pytest
 
 from location import CardinalPoint, Position
-from mars_rover import Rover
+from rover import Rover
 from tests.helpers import assert_rover_state
