+++ tests/units/test_mars_rover_left.py
@@ -1,7 +1,7 @@
 import pytest
 
 from location import CardinalPoint, Position
-from mars_rover import Rover
+from rover import Rover
 from tests.helpers import assert_rover_state
